move summary writing
